lazy loading, but also a bunch of refactoring -- probably shoulda done that in another branch
