+++ anvilfs/reference.py
@@ -1,28 +1,36 @@
 from io import BytesIO
 
-
-from .base import BaseAnVILFolder, BaseAnVILFile
+from .basefile import BaseAnVILFile
+from .basefolder import BaseAnVILFolder
 
 class ReferenceDataFolder(BaseAnVILFolder):
     def __init__(self, name, refs):
         super().__init__(name)
+        self.refs = refs
     
     def lazy_init(self):
-        self.init_references(refs)
+        self.init_references(self.refs)
 
     def init_references(self, refs):
         for source in refs:
             # source, e.g. hg38
-            source_baf = BaseAnVILFolder(source+"/")
-            self[source_baf.name] = source_baf
+            source_folder = RefereneDataSubfolder(source+"/")
+            self[source_folder.name] = source_folder
             # reftype, e.g. axiomPoly_resource_vcf
             for reftype in refs[source]:
-                reftype_baf = BaseAnVILFolder(reftype+"/")
-                source_baf[reftype_baf.name] = reftype_baf
+                reftype_folder = RefereneDataSubfolder(reftype+"/")
+                source_folder[reftype_folder.name] = reftype_folder
                 contents = ReferenceDataFile.make_rdfs(refs[source][reftype])
                 for c in contents:
-                    reftype_baf[c.name] = c
+                    reftype_folder[c.name] = c
 
+class RefereneDataSubfolder(BaseAnVILFolder):
+    def __init__(self, name):
+        super().__init__(name)
+        self.initialized = True
+    
+    def lazy_init(self):
+        pass
 
 class ReferenceDataFile(BaseAnVILFile):
     def __init__(self, blob):
